refactor: Price 입력을 훅으로 분리
