+++ src/pages/order/components/OrderPrice.tsx
@@ -1,5 +1,6 @@
-import React, {useRef, useState} from 'react'
+import React from 'react'
 import {formatNumber} from "../../../utils/formatNumber";
+import useIsEditing from "../hooks/useIsEditing";
 
 interface Props{
     price:string;
@@ -7,29 +8,7 @@ interface Props{
 }
 
 const OrderPrice: React.FC<Props> = ({price, setCurPrice}) => {
-    const [isEditing, setIsEditing] = useState(false);
-    const inputRef = useRef<HTMLInputElement>(null);
-
-    const handleKeyDown = (e:React.KeyboardEvent<HTMLInputElement>) => {
-        if (e.key === 'Enter') {
-            setIsEditing(false);
-        }
-    };
-
-    const handleClick = () => {
-        setIsEditing(true);
-        setTimeout(() => {
-            inputRef.current?.focus();
-        }, 0);
-    };
-
-    const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        const inputValue = e.target.value;
-
-        if (/^\d*$/.test(inputValue)) {
-            setCurPrice(Number(inputValue));
-        }
-    };
+    const {isEditing, handleKeyDown, handleClick, handleChange, handleBlur} = useIsEditing(setCurPrice)
 
     return (
         <div>
@@ -38,7 +17,7 @@ const OrderPrice: React.FC<Props> = ({price, setCurPrice}) => {
                     type="number"
                     value={price}
                     onChange={handleChange}
-                    onBlur={() => setIsEditing(false)}
+                    onBlur={handleBlur}
                     onKeyDown={handleKeyDown}
                     className={"text-3xl bg-gray-100 outline-none caret-indigo-600"}
                     autoFocus
